Restructured DataBase_Manager to separate script
