+++ getquote.py
@@ -2,11 +2,9 @@
 
 #* GetQuote.py                                   *#
 #* Nicholas DiBari                               *#
+#* --------------------------------------------- *#
 #* Prints quote from brainyquote.com from author *# 
 #* that user provides                            *#
-#* --------------------------------------------- *#
-#* Provides user interface for Get Quote         *#
-#* DataBase Management                           *#
 #* --------------------------------------------- *#
 
 import sys # Get argument
@@ -14,6 +12,8 @@ import requests # Get HTML
 import bs4 # Parsing HTML
 import shelve # Database Management
 from datetime import datetime # Date/Time Information
+
+from database import DataBase_Manager
 
 #1. Get name of author to search for
 def GetAuthor():
@@ -85,104 +85,6 @@ def SaveQuote(quotes_DB, quotes, author):
 
             except ValueError:
                 print('Enter in a number silly!')
-
-#5. Print Quotes
-def PrintDB(quotes_DB):
-    for key, value in quotes_DB.items():
-        print('{0} : '.format(key))
-        print(value)
-        print('-----------------------------------------------')
-
-#6. Delete Quote
-def DeleteQuote(quotes_DB): 
-    for key in quotes_DB.keys():
-        print(key)
-    
-    choice=input('Please select the title of the quote to delete: ')
-    
-    while choice not in list(quotes_DB.keys()):
-        choice=input('Input not found. Please enter the title of the quote to delete: ')
-        
-    #Get quote to delete 
-    quote = quotes_DB.get(choice)
-    
-    #Print quote to delete
-    print(choice + ": ")
-    print(quote)
-    print("-----------------------------------------------")
-    confirm = input('Are you sure you want to do delete this quote (y/n): ')
-
-    
-    #Delete quote from dictionary
-    if confirm == 'y' or confirm == 'Y':
-        del quotes_DB[choice]
-        print("Deleted\n")
-
-#7. Search Quote
-def SearchQuote(quotes_DB):
-    matches = [] # list of matched quotes
-    
-    while True:
-        #Get author to search for
-        to_search = input("Please enter an author to search for: ")
-    
-        #Go through list of authors, check if entered author is in any of the previous entries
-        for key in quotes_DB.keys():
-            if to_search in key:
-                matches.append(quotes_DB[key])
-
-        #If no matches, alert user no match was found
-        if not matches:
-            print("Sorry, did not find " + to_search + " in the database.")
-            
-        #Else, print matches
-        else:
-            print("Found the following quotes by " + to_search + ": ")
-            print(" ")
-            for quote in matches:
-                print(quote)
-                print("-----------------------------------------------")
-        
-        matches = [] # Reset list
-        choice = input("Would you like you search again? (y/n): ")
-        if choice == 'n' or choice == 'N':
-            break
-
-#8. Database Management
-def DataBase_Manager(quotes_DB):
-    flag = True
-    while flag:
-        print("Please enter a choice:")
-        print("1. Print all Quotes")
-        print("2. Delete a Quote")
-        print("3. Search for author")
-        print("4. [EXIT]")
-        choice = input()
-        
-        # ERROR CHECK
-        if int(choice)<1 or int(choice)>4:
-            print("Sorry that's not a valid choice. Try again")
-        
-        # PRINT QUOTES
-        elif int(choice) == 1:
-            PrintDB(quotes_DB)
-        
-        # DELETE QUOTE
-        elif int(choice) == 2:
-            DeleteQuote(quotes_DB)
-        
-        # SEARCH QUOTE
-        elif int(choice) == 3:
-            SearchQuote(quotes_DB)
-        
-        # [EXIT]
-        elif int(choice) == 4:
-            flag = False
-        
-        # ERROR CHECK
-        else:
-            print("Sorry that's not a valid input. Try again")
-        
     
 #Main Function
 def Main():
@@ -215,8 +117,6 @@ def Main():
     
     print("Goodbye!")   
     quotes_DB.close() # Close Database
-    
-    
 
 if __name__ == '__main__':
     Main()
